--- src/main/java/pl/edu/agh/io/umniedziala/activeApplicationMonitor/ActiveWindowNotFound.java
@@ -1,4 +1,0 @@
-package pl.edu.agh.io.umniedziala.activeApplicationMonitor;
-
-public class ActiveWindowNotFound extends Throwable {
-}
