+++ src/main/java/pl/edu/agh/io/umniedziala/monitors/activeApplicationMonitor/ActiveWindowNotFound.java
@@ -1,0 +1,4 @@
+package pl.edu.agh.io.umniedziala.monitors.activeApplicationMonitor;
+
+public class ActiveWindowNotFound extends Throwable {
+}
